add keep as as rclick option
